+++ Assets/Scripts/GunSetup.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class GunSetup : ScriptableObject
 {
     public float shotRate;
-    public int shotDamage;
+    public float shotDamage;
     public int shotDist;
 
     public GameObject GunModel;
